edge_device/src/streamlit_gcp/requirements.txt
streamlit==1.32.2
requests==2.31.0
google-cloud-firestore==2.14.0
pandas==2.2.1
Pillow==10.2.0
streamlit-autorefresh==0.0.3
